Translate some special CC values upon receipt
